+++ src/virtual-keyboard/virtual-keyboard.ts
@@ -794,6 +794,9 @@ export class VirtualKeyboard implements VirtualKeyboardInterface, EventTarget {
     const el = this._element;
     if (!el) return;
 
+    el.classList.toggle('is-math-mode', mf.mode === 'math');
+    el.classList.toggle('is-text-mode', mf.mode === 'text');
+
     el.classList.toggle('can-undo', mf.canUndo);
     el.classList.toggle('can-redo', mf.canRedo);
     el.classList.toggle('can-copy', !mf.selectionIsCollapsed);